UPDATE: 5-multithreads сделать правки ментора
