+++ homeworks/5-multithreads/calcNums.js
@@ -1,6 +1,8 @@
 const { performance } = require("perf_hooks");
 
 module.exports = (array) => {
+    performance.mark("start_calcNums");
+
     const nums = [];
     for (let i = 0; i < array.length; i++) {
         const num = array[i];
@@ -9,6 +11,7 @@ module.exports = (array) => {
             nums.push(num);
         }
     }
+    performance.measure("calcNums", "start_calcNums");
 
     return nums.length;
 };
